Pairs Violating BST code added

+++ Array/Count Alternating Subarrays.cpp
@@ -3,9 +3,9 @@
 class Solution {
 public:
     long long countAlternatingSubarrays(vector<int>& nums) {
-        int count = 0;
-        int start = 0;
-        for(int i = 0;i < nums.size();i++){
+        long long count = 0;
+        long long start = 0;
+        for(long long i = 0;i < nums.size();i++){
             if(i > 0 && nums[i] == nums[i-1]){
                 start = i;
             }
